@@ -1,10 +1,10 @@
-import { Component } from 'react';
+import { useEffect } from 'react';
 import Layout from '../components/Layout';
 
-let mountCount = 0;
+const Feedback = () => {
+  let mountCount = 0;
 
-export default class extends Component {
-  componentDidMount() {
+  useEffect(() => {
     if (++mountCount === 1) {
       const script = document.createElement('script');
 
@@ -20,22 +20,21 @@ export default class extends Component {
       basePath: '/feedback',
       ssoToken: null
     });
-  }
+  }, []);
 
-  render() {
-    return (
-      <Layout title="Feedback" sidebarActiveItemId="feedback">
-        <div className="feedback" data-canny />
-        <style jsx>{`
-          .feedback {
-            flex-grow: 1;
-            overflow: auto;
+  return (
+    <Layout title="Feedback" sidebarActiveItemId="feedback">
+      <div className="feedback" data-canny />
+      <style jsx>{`
+        .feedback {
+          flex-grow: 1;
+          overflow: auto;
+          padding: 20px 30px;
+          background-color: #fff;
+        }
+      `}</style>
+    </Layout>
+  );
+};
 
-            padding: 20px 30px;
-            background-color: #fff;
-          }
-        `}</style>
-      </Layout>
-    );
-  }
-}
+export default Feedback;
